requirements.txt
boto3==1.42.17
capsolver==1.0.7
discord.py==2.6.3
jpholiday==1.0.3
matplotlib==3.10.8
numpy==2.4.4
pandas==2.3.2
playwright==1.58.0
playwright-recaptcha==0.5.1
playwright-stealth==2.0.3
python-dotenv==1.1.1
requests==2.33.1
seaborn==0.13.2
